requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-storages==1.14.6
boto3==1.37.37
django-filter==25.1
pydantic==2.11.7
geopy==2.4.1
openpyxl==3.1.5
